Added CommunicationError and InvalidCoordinates Exception

+++ GestioneIncendi/src/main/java/salvomercurio/GestioneIncendi/interfaces/FireAlarm.java
@@ -2,6 +2,6 @@ package salvomercurio.GestioneIncendi.interfaces;
 
 public interface FireAlarm {
 
-	void triggerAlarm(Long id, double latitude, double longitude, double smokeLevel);
+	void triggerAlarm(int id, double latitude, double longitude, double smokeLevel);
 
 }
