+++ GestioneIncendi/src/main/java/salvomercurio/GestioneIncendi/util/FireAlarmLevelImpl.java
@@ -5,7 +5,7 @@ import salvomercurio.GestioneIncendi.interfaces.FireAlarm;
 public class FireAlarmLevelImpl implements FireAlarm {
 
 	@Override
-	public void triggerAlarm(Long id, double latitude, double longitude, double smokeLevel) {
+	public void triggerAlarm(int id, double latitude, double longitude, double smokeLevel) {
 		String url = "http://host/alarm?idsonda=" + id + "&lat=" + latitude + "&lon=" + longitude + "&smokelevel="
 				+ smokeLevel;
 	}
